Changed component name to generatepdf to be more descriptive
